Enum kanidm_unix::KanidmUnixOpt

pub enum KanidmUnixOpt {
    AuthTest {
        debug: bool,
        account_id: String,
    },
    CacheClear {
        debug: bool,
        really: bool,
    },
    CacheInvalidate {
        debug: bool,
    },
    Status {
        debug: bool,
    },
    Version {
        debug: bool,
    },
}

Variants

AuthTest

Fields

debug: bool

account_id: String

Test authentication of a user via the unixd resolver “pam” channel. This does not test that your pam configuration is correct - only that unixd is correctly processing and validating authentications.

CacheClear

Fields

debug: bool

really: bool

Erase the content of the unixd resolver cache. You should probably use invalidate instead.

CacheInvalidate

Fields

debug: bool

Invalidate, but don’t erase the content of the unixd resolver cache. This will force the unixd daemon to refresh all user and group content immediately. If the connection is offline, entries will still be available and will be refreshed as soon as the daemon is online again.

Status

Fields

debug: bool

Check that the unixd daemon is online and able to connect correctly to the kanidmd server.

Version

Fields

debug: bool

Show the version of this tool.